change: move OutputView to the Tools namespace
