+++ src/Beutl/Views/Dialogs/OutputProgressDialog.axaml.cs
@@ -1,4 +1,5 @@
 ﻿using Beutl.ViewModels;
+using Beutl.ViewModels.Tools;
 using FluentAvalonia.UI.Controls;
 
 namespace Beutl.Views.Dialogs;
