+++ src/Beutl/Views/Tools/OutputView.axaml
@@ -1,12 +1,11 @@
-<UserControl x:Class="Beutl.Views.OutputView"
+<UserControl x:Class="Beutl.Views.Tools.OutputView"
              xmlns="https://github.com/avaloniaui"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-             xmlns:icons="using:FluentIcons.FluentAvalonia"
              xmlns:lang="using:Beutl.Language"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
              xmlns:tools="using:Beutl.Views.Tools"
-             xmlns:viewModel="using:Beutl.ViewModels"
+             xmlns:viewModel="using:Beutl.ViewModels.Tools"
              d:DesignHeight="800"
              d:DesignWidth="600"
              x:CompileBindings="True"
