+++ src/Beutl/Views/Tools/OutputView.axaml.cs
@@ -3,9 +3,10 @@ using Avalonia.Interactivity;
 using Avalonia.Platform.Storage;
 
 using Beutl.ViewModels;
+using Beutl.ViewModels.Tools;
 using Beutl.Views.Dialogs;
 
-namespace Beutl.Views;
+namespace Beutl.Views.Tools;
 
 public partial class OutputView : UserControl
 {
